Fixed installer.
Found out that the pip install was not installing the mideu config file which is used as a fallback when no config file is provided. This has now been corrected. I still don't really get the whole installing files outside of the standard process.. the doco in this regard is thin.. Is working now. So you can now do a `pip install mciutil` and the utility will work out of the box.
Changed to use pkg_resources package to get the location of the package rather than using __name__ etc..